+++ CHANGELOG.md
@@ -15,10 +15,13 @@ This release extends support to OneView Rest API version 1800 (OneView v5.30) fo
 - Network Set
 - Scope
 - Server Certificate
+- Server Hardware
+- Server Hardware Type
+- Server Profile
+- Server Profile Template
 - Storage Pool
 - Storage System
 - Uplink Set
-
 
 ## v5.13.0